gpu: move the decline undo machinery into its own header
Everything that lets a declined loop be handed back exactly as it was
found -- the body write-back, the splice snapshot, the saved scope-local
array types, and the three RAII guards that give a gather, a kernel
draft and a splice back on every early exit -- is one idea, and the
visitor is its only caller.

The three types that were in an anonymous namespace become ordinary
namespace-scope types, since a header cannot carry an anonymous
namespace. Nothing else about them changes.
